+++ app/controllers/ClientLogin.java
@@ -50,14 +50,11 @@ public class ClientLogin extends Controller
 
 		public String validate()
 		{
-			Logger.info("validating id: " + id);			
-			//Logger.info("validating password: " + password);
-			Logger.info("validating experimentId: " + experimentId);			
-			Logger.info("validating experimentInstanceId: " + experimentInstanceId);			
-			/*
-			 * TODO: Hash function here
-			 */
-			//return "Invalid user or password";
+			Logger.info("Client Login, ID: " + id);
+
+			if (! ScriptBoard.checkPassword(password)) {
+				return "Invalid user or password";
+			}
 			return null;
 		}
 	}